For consistency, reput the loop over real arrays. Should be changed later

+++ src/Atoms/barklem.f90
@@ -137,7 +137,7 @@ MODULE barklem
    logical, intent(out) :: res
    logical :: determined
    real(8) :: Si, Sj
-   integer :: Jj, Ji
+   real(8) :: Jj, Ji
    real(8) :: neff1, neff2, nefftmp
    real(8) :: findex1, findex2, reducemass, meanvelocity
    real(8) :: crossmean, E_Rydberg2, deltaEi, deltaEj
